// frontend/bindings/github.com/go-ctap/kit/model/safety/models.ts
// Cynhyrchwyd y ffeil hon yn awtomatig. PEIDIWCH Â MODIWL
// This file is automatically generated. DO NOT EDIT

// eslint-disable-next-line @typescript-eslint/ban-ts-comment
// @ts-ignore: Unused imports
import { Create as $Create } from "@wailsio/runtime";

export enum PreviewMode {
    /**
     * The Go zero value for the underlying type of the enum.
     */
    $zero = "",

    PreviewModeDryRun = "dry-run",
    PreviewModeExecute = "execute",
};

export enum Severity {
    /**
     * The Go zero value for the underlying type of the enum.
     */
    $zero = "",

    SeverityInfo = "info",

    SeverityWarning = "warning",

    SeverityDestructive = "destructive",
};

export class Warning {
    "severity": Severity;
    "code": string;
    "message": string;

    /** Creates a new Warning instance. */
    constructor($$source: Partial<Warning> = {}) {
        if (!("severity" in $$source)) {
            this["severity"] = Severity.$zero;
        }
        if (!("code" in $$source)) {
            this["code"] = "";
        }
        if (!("message" in $$source)) {
            this["message"] = "";
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new Warning instance from a string or object.
     */
    static createFrom($$source: any = {}): Warning {
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        return new Warning($$parsedSource as Partial<Warning>);
    }
}